requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==0.5.0
annotated-types==0.7.0
art==6.4
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==1.5.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pycm==3.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.12.12
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.19.1
Werkzeug==3.1.3
yarl==1.18.3
